+++ src/main/java/com/enext/edu/registration/model/Program.java
@@ -1,14 +1,14 @@
 package com.enext.edu.registration.model;
 
 public class Program {
-    private Long id;
+    private Integer id;
     private String name;
 
-    public Program(Long id, String name) {
+    public Program(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
     public String getName() { return name; }    
 }
