+++ src/main/java/com/enext/edu/registration/repository/CourseBySemesterRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CourseBySemesterRepository extends JpaRepository<CourseBySemester, Long> {
-    List<CourseBySemester> findBySemesterId(Long semesterId);
+public interface CourseBySemesterRepository extends JpaRepository<CourseBySemester, Integer> {
+    List<CourseBySemester> findBySemesterId(Integer semesterId);
 }
