+++ src/main/java/com/enext/edu/registration/repository/StudentRegistrationRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StudentRegistrationRepository extends JpaRepository<StudentRegistration, Long> {
+public interface StudentRegistrationRepository extends JpaRepository<StudentRegistration, Integer> {
 
-    List<StudentRegistration> findByStudentId(Long studentId);
+    List<StudentRegistration> findByStudentId(Integer studentId);
 
-    StudentRegistration findByStudentIdAndSemesterId(Long studentId, Long semesterId);
+    StudentRegistration findByStudentIdAndSemesterId(Integer studentId, Integer semesterId);
 }
